added ed25519 key generation (dropbearkey -t ed25519)
